+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 0.10.0
+
+### New features
+
+- Feature "alloc"
+    - Required for `many_to_vec` and `many_to_map_ordered` (no requirement on "std")
+- Add extension functions for starting a parse. Instead of writing
+  e.g. `parse(p, input)` you can write `p.parse(input)`.
+- Made all parsers and combinators `const`. Note that parsing is _not_ `const` (yet).
+- Made Utf8Whitespace implement Prefix.
+- New parser `fold_state`, that uses the user state instead of a provided accumulator.
+- `many_to_array` now supports `BorrowMut<[T, N]>` in initializer.
+
 ## 0.9.0
 
 ### New features